Product Management Application System
The system provided is a Python script that utilizes the PyODBC library for connecting to a SQLite database and the Tkinter library in order to create a graphical user interface (GUI). Check out a description of the code:

The necessary modules are imported: pyodbc for database connectivity and tkinter for GUI development.
The verifica_credenciais() function is defined. It establishes a connection to the SQLite database, executes a query to check if the entered username and password exist in the database, and retrieves the user data if found.
If the user exists, a new window is created for product registration. It includes input fields for product name, description, and price.
The cadastrar() function is defined, which is called when the "Novo" button is clicked. It creates a new window for product registration and includes input fields for product name, description, and price. There are also buttons for saving and canceling the registration.
The listar_dados() function is defined to fetch the product data from the database and populate it in a treeview widget in the main window.
The editar_dados(event) function is called when an item in the treeview is selected. It creates a new window for editing the selected product's details, with input fields pre-filled with the existing values. It also includes buttons for saving the changes or deleting the product.
The code includes the configuration of the treeview widget to display the product data in columns (ID, Product Name, Description, and Price).
The script sets up the GUI elements and starts the main event loop.
Overall, the code establishes a connection to a SQLite database, allows users to log in, and provides functionality for adding, editing, and deleting product records through a graphical user interface.
If you want to test it and log in, you can use as login: "Gisélia". Password: "290552"